moved isTrackingAllowed to helpers
