+++ packages/cms-lib/lib/config/file.js
@@ -188,14 +188,6 @@ const deleteEmptyConfigFile = () => {
   );
 };
 
-const isTrackingAllowed = () => {
-  if (!configFileExists() || configFileIsBlank()) {
-    return true;
-  }
-  const { allowUsageTracking } = getAndLoadConfigIfNeeded();
-  return allowUsageTracking !== false;
-};
-
 module.exports = {
   createEmptyConfigFile,
   configFileExists,
@@ -204,7 +196,6 @@ module.exports = {
   getAndLoadConfigIfNeeded,
   getConfig,
   getConfigPath,
-  isTrackingAllowed,
   setConfig,
   setDefaultConfigPathIfUnset,
   loadConfig,
