+++ packages/cms-lib/lib/config/helpers.js
@@ -1,4 +1,17 @@
+const {
+  configFileExists,
+  configFileIsBlank,
+  getAndLoadConfigIfNeeded,
+} = require('./file');
 const { Mode } = require('../constants');
+
+const isTrackingAllowed = () => {
+  if (!configFileExists() || configFileIsBlank()) {
+    return true;
+  }
+  const { allowUsageTracking } = getAndLoadConfigIfNeeded();
+  return allowUsageTracking !== false;
+};
 
 /**
  * Returns mode(draft/publish) or undefined if invalid mode
@@ -10,4 +23,5 @@ const getMode = mode => {
 
 module.exports = {
   getMode,
+  isTrackingAllowed,
 };
